op-node/rollup: move startPayload under EngineController
